organized tests to their respective parts of application
Tests for the app + db were moved to the integrate/hms_smoke app folder.
